refactor(utils): remove CheckpointResult and use BaseCTCModule directly
