+++ src/thunder/citrinet/compatibility.py
@@ -19,10 +19,11 @@ from torchaudio.datasets.utils import extract_archive
 
 from thunder.blocks import conv1d_decoder
 from thunder.citrinet.blocks import CitrinetEncoder
+from thunder.module import BaseCTCModule
 from thunder.quartznet.compatibility import load_quartznet_weights
 from thunder.quartznet.transform import FilterbankFeatures
 from thunder.text_processing.transform import BatchTextTransformer
-from thunder.utils import BaseCheckpoint, CheckpointResult, download_checkpoint
+from thunder.utils import BaseCheckpoint, download_checkpoint
 
 
 # fmt:off
@@ -115,7 +116,7 @@ def fix_vocab(vocab_tokens: List[str]) -> List[str]:
 
 def load_citrinet_checkpoint(
     checkpoint: Union[str, CitrinetCheckpoint], save_folder: str = None
-) -> CheckpointResult:
+) -> BaseCTCModule:
     """Load from the original nemo checkpoint.
 
     Args:
@@ -145,10 +146,11 @@ def load_citrinet_checkpoint(
 
         weights_path = extract_path / "model_weights.ckpt"
         load_quartznet_weights(encoder, decoder, str(weights_path))
-        return CheckpointResult(
+        module = BaseCTCModule(
             encoder,
             decoder,
             audio_transform,
             text_transform,
             encoder_final_dimension=640,
         )
+        return module.eval()
